requirements.txt
Flask
gunicorn
python-dotenv
Flask-SQLAlchemy
Werkzeug
python-binance
requests
psycopg2-binary
